stream/producer/requirements.txt
kafka-python
requests
protobuf==3.11.0
yliveticker
pandas
yfinance